Feature + Change
Animdl package: Works in the terminal.
GDLauncher package: Works in terminal but can be added to a desktop file.
XClicker package: Works in terminal but can be added to a desktop file.

Will need to build the last 2 from source and add them to the desktop files because the appimage wrappers cant do that afaik.