Picom test
Supposedly someone was having issues with the xrender backend causing OOM panics on the Xorg process. Maybe it's the same case with mine? Whatever issues I had when originally using the glx backend seem to be gone (or I haven't noticed them yet) so I'll try using glx for a while and see.